refactor(app): break runtime import cycles in src/app core
Two runtime cycle edges cut:

- i18n/index.ts <-> app/constants.ts: LANGUAGE_PREF_KEY now lives in i18n
  (where it's used); constants re-exports it. Kills the genuine runtime
  cycle i18n -> constants -> i18n.
- lib/{openwork-server,opencode,den,den-telemetry}.ts no longer import the
  1,199-line utils barrel (which drags in i18n + constants) just for
  isDesktopRuntime: moved isElectronRuntime/isDesktopRuntime to new leaf
  module lib/runtime-env.ts; utils re-exports for the 7 react-app callers.

madge --circular: 7 cycles -> 2, and both remaining close via import-type
edges only (extensions->types, den-session-events->den), i.e. zero runtime
cycles. Those type edges dissolve in the upcoming shared-contract step.

Tested:
- pnpm typecheck (clean)
- bun test tests/ — 51 pass, 2 fail: verified identical failures on base
  via git stash (stale tests referencing removed exports, pre-existing)
- pnpm build (vite production build clean)
- live: vite dev server boots, / -> 200, transformed i18n/constants modules
  serve with the new export graph
